+++ cmd/issue.go
@@ -9,6 +9,7 @@ import (
 )
 
 var (
+	issueRobot       string
 	issueMetafile    string
 	issueAttachments []string
 )
@@ -21,7 +22,7 @@ var issueCmd = &cobra.Command{
 		if common.DebugFlag {
 			defer common.Stopwatch("Feedback issue lasted").Report()
 		}
-		err := operations.ReportIssue(issueMetafile, issueAttachments, dryFlag)
+		err := operations.ReportIssue(issueRobot, issueMetafile, issueAttachments, dryFlag)
 		if err != nil {
 			pretty.Exit(1, "Error: %s", err)
 		}
@@ -34,4 +35,5 @@ func init() {
 	issueCmd.Flags().StringVarP(&issueMetafile, "report", "r", "", "Report file in JSON form containing actual issue report details.")
 	issueCmd.Flags().StringArrayVarP(&issueAttachments, "attachments", "a", []string{}, "Files to attach to issue report.")
 	issueCmd.Flags().BoolVarP(&dryFlag, "dryrun", "d", false, "Don't send issue report, just show what would report be.")
+	issueCmd.Flags().StringVarP(&issueRobot, "robot", "", "", "Full path to 'robot.yaml' configuration file. [optional]")
 }